requirements.txt
requests==2.32.3
beautifulsoup4==4.13.3
openai==1.70.0
python-dotenv==1.1.0
tqdm==4.67.1
pandas==2.2.3
